requirements.txt
attrs==25.4.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
referencing==0.37.0
regex==2025.10.23
rpds-py==0.27.1
ruff==0.14.1